mcni/components/NeutronStorage: fixed a few bugs and added another test case.
git-svn-id: svn://danse.us/MCViNE/trunk@178 bd03dbb6-ca95-436e-b6ad-03a95796df24

+++ packages/mcni/mcni/components/_neutron_storage_impl/PacketSizeStore.py
@@ -24,7 +24,11 @@ class PacketSizeStore:
 
     def __init__(self, path):
         self.path = path
-        self.handle = open(path, 'w+')
+        import os
+        if os.path.exists( path ):
+            self.handle = open(path, 'r')
+        else:
+            self.handle = open(path, 'w+')
         self.size = None
         return
 
@@ -34,7 +38,7 @@ class PacketSizeStore:
             text = self.handle.read()
             if len(text) == 0: return 
             size = long( text )
-            self.size = long( self.handle.read() )
+            self.size = size
         return self.size
